Split LocalExecutor and RemoteExecutor to different modules
